requirements-dev.txt
annotated-types==0.7.0
backports.tarfile==1.2.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
docstring_parser==0.16
docutils==0.21.2
exceptiongroup==1.2.2
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
joblib==1.4.2
keyring==25.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
mypy-extensions==1.0.0
nh3==0.2.18
numpy==2.1.3
nvidia-nccl-cu12==2.23.4
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
shapely==2.0.6
six==1.16.0
threadpoolctl==3.5.0
tomli==2.0.2
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xgboost==2.1.2
zipp==3.20.2
